+++ lab-06-orm/src/main/java/com/cydeo/lab06orm/entity/Payment.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Data
 @Entity
@@ -15,7 +16,7 @@ public class Payment {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Double paidPrice;
+    private BigDecimal paidPrice;
     @Enumerated(EnumType.STRING)
     private PaymentMethod paymentMethod;
